標準入力のリダイレクトを実装
